+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+## [v0.205.0]
+
 ### Added
 
 - `org`: `Address` now includes a `state` code, for countries that require them.